Honor those who have been here before thee !
All thou didst was rewrite the manpage to make no sense, Vibhore. I have edited it back to appropriately mention the lineage of the manpage (ellenor@umbrellix.net) as well as the ircd (aspircd is a fork of irca, asterIRC's ircd).

What the hell was that with ircd-nine anyway? Asinine poor grammar and factual incorrectness. Didst thou mandoc|less the page before thou shippedest it?